Pass the sample_range_limit instead of the whole cinfo to the idct functions
The only thing out of the cinfo that the idct functions ever use
is the sample_range_limit, or its subset IDCT_range_limit(cinfo).

We split that macro into two:

This is so that the existing code which already uses IDCT_range_limit() can remain unchanged.

We pass the full array sample_range_limit to the idct functions.
Most of them use the subset and extract it with the first macro,
but jpeg_idct_float uses the full array.

This removes the last use of cinfo in the simd/ directory.